requirements.txt
numpy==1.24.1
pdfkit==1.0.0
pyyaml==6.0
